Update play script

+++ play.py
@@ -1,7 +1,8 @@
 from __future__ import print_function
 from neat import Checkpointer
-from neat.nn import FeedForwardNetwork
+from neat.nn import FeedForwardNetwork, RecurrentNetwork
 import sys
+import os
 
 import gzip
 try:
@@ -69,6 +70,7 @@ def simulateGame(net=None):
     ai_fitness = 0
     training_filename = "training.csv"
     if net is not None:
+        last_change = 0
         while not game.isFinished():
             output = net.activate(game.state())
             if rounding:
@@ -89,45 +91,56 @@ def simulateGame(net=None):
         training_filename = "training-{}.csv".format(ai_fitness)
         print("AI Fitness: ", ai_fitness)
 
-    with open(training_filename, "w") as f:
-        while f.tell() == 0: # Haven't written anything to train
-            last_change = 0
-            fitness = 0
-            while not game.isFinished():
-                game.render()
-                if pressed.count(0.) == len(pressed) and net is not None:
-                    output = net.activate(game.state())
-                    if rounding:
-                        output = [round(val) for val in output]
-                    game.step(output)
+    answer = input("Create training file {}? (y/N)".format(training_filename))
+    if answer.upper() == "Y" and os.path.isfile(training_filename):
+        answer = input("It already exists, are you sure? (Y/N)")
 
-                    if game.fitness() > fitness:
-                        fitness = game.fitness()
-                        last_change = 0
-                    else:
-                        last_change += 1
-                        if last_change > 250: # Around 10seconds in game
-                            break
-                        if last_change > 125: # I guess around 5 seconds in game?
-                            break
-                else:
-                    if isFocussed():
-                        f.write('"' + '","'.join(map(str, list(game.state()) + pressed)) + '"\n')
-                        game.step(pressed)
-                        time.sleep(0.0175)
+    if answer.upper() == "Y":
+        with open(training_filename, "w") as f:
+            while f.tell() == 0: # Haven't written anything to train
+                run(game, net, f)
 
-            if f.tell() == 0:
-                print("Haven't got anything to train against yet")
-                ai_fitness = fitness
-                game.reset()
-            elif game.fitness() <= ai_fitness:
-                print("Haven't actually improved yet...")
-                f.seek(0, 0)
-                game.reset()
+                if f.tell() == 0:
+                    print("Haven't got anything to train against yet")
+                    game.reset()
+                elif game.fitness() <= ai_fitness:
+                    print("Haven't actually improved yet...")
+                    f.seek(0, 0)
+                    game.reset()
+    else:
+        run(game, net)
 
     game.close()
 
     return game.fitness()
+
+def run(game, net, training_file=None):
+    fitness = 0
+    last_change = 0
+
+    while not game.isFinished():
+        game.render()
+        if pressed.count(0.) == len(pressed) and net is not None:
+            output = net.activate(game.state())
+            if rounding:
+                output = [round(val) for val in output]
+            game.step(output)
+
+            if game.fitness() > fitness:
+                fitness = game.fitness()
+                last_change = 0
+            else:
+                last_change += 1
+                if last_change > 250: # Around 10seconds in game
+                    break
+                if last_change > 125: # I guess around 5 seconds in game?
+                    break
+        else:
+            if isFocussed():
+                if training_file is not None:
+                    training_file.write('"' + '","'.join(map(str, list(game.state()) + pressed)) + '"\n')
+                game.step(pressed)
+                time.sleep(0.0175)
 
 if __name__ == '__main__':
     # TODO parse sysargs properly
@@ -135,7 +148,10 @@ if __name__ == '__main__':
     if len(sys.argv) >= 3:
         p = Checkpointer.restore_checkpoint(sys.argv[1])
         genome = p.population[int(sys.argv[2])]
-        net = FeedForwardNetwork.create(genome, p.config)
+        if p.config.genome_config.feed_forward:
+            net = FeedForwardNetwork.create(genome, p.config)
+        else:
+            net = RecurrentNetwork.create(genome, p.config)
     elif len(sys.argv) >= 2:
         with gzip.open(sys.argv[1]) as f:
             net = pickle.load(f)
